Code refactor - vue page update and order change
