+++ src/router/index.js
@@ -5,7 +5,6 @@ import PatientInfo from "../views/PatientInfo.vue";
 import UploadDocuments from "../views/UploadDocuments.vue";
 import ReviewPage from "../views/ReviewPage.vue";
 import SubmissionPage from "../views/SubmissionPage.vue";
-import PractitionerUploadTool from "../views/PractitionerUploadTool.vue";
 
 export const routes = {
   GET_STARTED: {
@@ -19,12 +18,6 @@ export const routes = {
     title: "Practitioner Info",
     name: "PractitionerInfo",
     component: PractitionerInfo,
-  },
-  PRACTITIONER_UPLOAD_TOOL: {
-    path: "/practitioner-upload-tool",
-    title: "Practitioner Upload Tool",
-    name: "PractitionerUploadTool",
-    component: PractitionerUploadTool,
   },
   PATIENT_INFO: {
     path: "/patient-info",
@@ -55,7 +48,6 @@ export const routes = {
 export const stepRoutes = [
   { ...routes.GET_STARTED },
   { ...routes.PRACTITIONER_INFO },
-  { ...routes.PRACTITIONER_UPLOAD_TOOL },
   { ...routes.PATIENT_INFO },
   { ...routes.UPLOAD_DOCUMENTS },
   { ...routes.REVIEW_PAGE },
@@ -65,7 +57,6 @@ export const stepRoutes = [
 export const routeStepOrder = [
   routes.GET_STARTED,
   routes.PRACTITIONER_INFO,
-  routes.PRACTITIONER_UPLOAD_TOOL,
   routes.PATIENT_INFO,
   routes.UPLOAD_DOCUMENTS,
   routes.REVIEW_PAGE,
@@ -97,11 +88,6 @@ const router = createRouter({
       path: "/practitioner-info",
       name: "PractitionerInfo",
       component: () => import("../views/PractitionerInfo.vue"),
-    },
-    {
-      path: "/practitioner-upload-tool",
-      name: "PractitionerUploadTool",
-      component: () => import("../views/PractitionerUploadTool.vue"),
     },
     {
       path: "/patient-info",
